fixed pathing error: moved utils dir into server/src. edited user model/schema to use comparepassword method
