+++ server/src/models/User.ts
@@ -6,6 +6,11 @@ interface IUser extends Document {
 	email: string;
 	password: string;
 	createdAt: Date;
+	comparePassword(candidatePassword: string): Promise<boolean>;
 }
+UserSchema.methods.comparePassword = async function (candidatePassword: string) {
+	const user = this as IUser;
+	return candidatePassword === user.password;
+};
 
 export default mongoose.model<IUser>("User", UserSchema);
