+++ server/src/schemas/UserSchema.ts
@@ -6,4 +6,6 @@ const UserSchema: Schema = new Schema({
     password: { type: String, required: true }
 });
 
+
+
 export { UserSchema };
